Agregué assets de tarjetas de Brasil

+++ sdk/src/main/java/com/mercadopago/FrontCardActivity.java
@@ -160,7 +160,7 @@ public abstract class FrontCardActivity extends MercadoPagoActivity implements C
             return s.toString();
         }
         try {
-            Integer value = Integer.valueOf(s.toString());
+            Long value = Long.valueOf(s.toString());
             Locale.setDefault(Locale.GERMAN);
             return String.format("%,d", value);
         } catch (NumberFormatException e) {
